sahayak-coordinator/sub_agents/sahayak_worksheet_agent/requirements.txt
# Core dependencies
google-adk
google-cloud-firestore
google-cloud-storage
google-cloud-vision
google-generativeai
google-auth
python-dotenv

# Document processing dependencies
pdfplumber>=0.7.0  # For PDF text extraction
python-docx>=0.8.11  # For Word document processing

# File generation dependencies
python-pptx>=0.6.21  # For PowerPoint generation
reportlab>=3.6.0  # For PDF worksheet generation

# Additional dependencies for document processing
Pillow>=9.0.0  # For image handling and Vision API
pandas>=1.3.0  # For data processing if needed